+++ src/components/sdk/MeetingControls/AudioInputControl.tsx
@@ -3,10 +3,10 @@
 
 import React, { useEffect, useState } from 'react';
 
-import useSelectAudioInputDevice from '../../../hooks/sdk/useSelectAudioInputDevice';
 import { useToggleLocalMute } from '../../../hooks/sdk/useToggleLocalMute';
 import { useAudioInputs } from '../../../providers/DevicesProvider';
 import { useLogger } from '../../../providers/LoggerProvider';
+import { useMeetingManager } from '../../../providers/MeetingProvider';
 import { isOptionActive } from '../../../utils/device-utils';
 import { ControlBarButton } from '../../ui/ControlBar/ControlBarButton';
 import { Microphone } from '../../ui/icons';
@@ -32,7 +32,7 @@ const AudioInputControl: React.FC<Props> = ({
   ...rest
 }) => {
   const logger = useLogger();
-  const selectAudioInput = useSelectAudioInputDevice();
+  const meetingManager = useMeetingManager();
   const { muted, toggleMute } = useToggleLocalMute();
   const { devices, selectedDevice } = useAudioInputs();
   const [dropdownOptions, setDropdownOptions] = useState<PopOverItemProps[]>(
@@ -42,7 +42,7 @@ const AudioInputControl: React.FC<Props> = ({
   useEffect(() => {
     const handleClick = async (deviceId: string): Promise<void> => {
       try {
-        await selectAudioInput(deviceId);
+        await meetingManager.startAudioInputDevice(deviceId);
       } catch (error) {
         logger.error('AudioInputControl failed to select audio input device');
       }
@@ -60,7 +60,12 @@ const AudioInputControl: React.FC<Props> = ({
     };
 
     getDropdownOptions();
-  }, [devices, selectedDevice, selectAudioInput]);
+  }, [
+    devices,
+    selectedDevice,
+    meetingManager,
+    meetingManager.startAudioInputDevice,
+  ]);
 
   return (
     <ControlBarButton